requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
auth0-python==4.8.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cohere==5.13.12
colorama==0.4.6
coloredlogs==15.0.1
crewai==0.102.0
crewai-tools==0.36.0
cryptography==44.0.1
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
durationpy==0.9
embedchain==0.1.126
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.8
fastavro==1.10.0
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.80.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
gptcache==0.1.44
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
instructor==1.7.2
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
json5==0.10.0
json_repair==0.36.0
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.0
lancedb==0.19.0
langchain==0.3.18
langchain-cohere==0.3.5
langchain-community==0.3.17
langchain-core==0.3.35
langchain-experimental==0.3.4
langchain-google-genai==2.0.9
langchain-openai==0.2.14
langchain-text-splitters==0.3.6
langgraph==0.3.2
langgraph-checkpoint==2.0.16
langgraph-prebuilt==0.1.1
langgraph-sdk==0.1.53
langsmith==0.1.147
litellm==1.60.2
load-dotenv==0.1.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.48
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.62.0
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
portalocker==2.10.1
posthog==3.13.0
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pylance==0.23.0
pyparsing==3.2.1
pypdf==5.3.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.394
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
pywin32==308
PyYAML==6.0.2
qdrant-client==1.13.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
schema==0.7.7
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.7.0
tokenizers==0.20.3
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uv==0.5.31
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
